+++ src/main/java/io/github/alexengrig/lambdax/collection/MapX.java
@@ -37,19 +37,35 @@ public final class MapX {
     }
 
     /**
+     * <p>Returns the carrying of {@link java.util.Map#containsKey(Object)} negative:
+     * key -&gt; map -&gt; !map.containsKey(key).</p>
+     *
+     * @param key a key of {@link K} that is passed as the argument to {@link java.util.Map#containsKey(Object)}
+     * @param <K> a type of keys in a map
+     * @param <V> a type of values in a map
+     * @return a {@link java.util.function.Predicate}
+     * @see java.util.Map#containsKey(Object)
+     * @see java.util.function.Predicate
+     * @since 0.1.2
+     */
+    public static <K, V> Predicate<Map<? extends K, ? extends V>> containsKey(K key) {
+        return map -> map.containsKey(key);
+    }
+
+    /**
      * <p>Returns the carrying of {@link java.util.Map#containsKey(Object)}:
      * key -&gt; map -&gt; map.containsKey(key).</p>
      *
      * @param key a key of {@link K} that is passed as the argument to {@link java.util.Map#containsKey(Object)}
-     * @param <K> a type of keys in a map.
-     * @param <V> a type of values in a map.
-     * @return a {@link java.util.function.Predicate} .
+     * @param <K> a type of keys in a map
+     * @param <V> a type of values in a map
+     * @return a {@link java.util.function.Predicate}
      * @see java.util.Map#containsKey(Object)
      * @see java.util.function.Predicate
      * @since 0.1.0
      */
-    public static <K, V> Predicate<Map<? extends K, ? extends V>> containsKey(K key) {
-        return map -> map.containsKey(key);
+    public static <K, V> Predicate<Map<? extends K, ? extends V>> notContainsKey(K key) {
+        return map -> !map.containsKey(key);
     }
 
     /**
@@ -57,9 +73,9 @@ public final class MapX {
      * value -&gt; map -&gt; map.containsValue(value).</p>
      *
      * @param value a value of {@link V} that is passed as the argument to {@link java.util.Map#containsValue(Object)}
-     * @param <K>   a type of keys in a map.
-     * @param <V>   a type of values in a map.
-     * @return a {@link java.util.function.Predicate} .
+     * @param <K>   a type of keys in a map
+     * @param <V>   a type of values in a map
+     * @return a {@link java.util.function.Predicate}
      * @see java.util.Map#containsValue(Object)
      * @see java.util.function.Predicate
      * @since 0.1.0
@@ -69,13 +85,29 @@ public final class MapX {
     }
 
     /**
+     * <p>Returns the carrying of {@link java.util.Map#containsValue(Object)} negative:
+     * value -&gt; map -&gt; !map.containsValue(value).</p>
+     *
+     * @param value a value of {@link V} that is passed as the argument to {@link java.util.Map#containsValue(Object)}
+     * @param <K>   a type of keys in a map
+     * @param <V>   a type of values in a map
+     * @return a {@link java.util.function.Predicate}
+     * @see java.util.Map#containsValue(Object)
+     * @see java.util.function.Predicate
+     * @since 0.1.0
+     */
+    public static <K, V> Predicate<Map<? extends K, ? extends V>> notConntainsValue(V value) {
+        return map -> !map.containsValue(value);
+    }
+
+    /**
      * <p>Returns the carrying of {@link java.util.Map#get(Object)}:
      * key -&gt; map -&gt; map.get(key).</p>
      *
      * @param key a key of {@link K} that is passed as the argument to {@link java.util.Map#get(Object)}
-     * @param <K> a type of keys in a map.
-     * @param <V> a type of values in a map.
-     * @return a {@link java.util.function.Function} .
+     * @param <K> a type of keys in a map
+     * @param <V> a type of values in a map
+     * @return a {@link java.util.function.Function}
      * @see java.util.Map#get(Object)
      * @see java.util.function.Function
      * @since 0.1.0
@@ -90,9 +122,9 @@ public final class MapX {
      *
      * @param key   a key of {@link K} that is passed as the first argument to {@link java.util.Map#put(Object, Object)}
      * @param value a value of {@link V} that is passed as the second argument to {@link java.util.Map#put(Object, Object)}
-     * @param <K>   a type of keys in a map.
-     * @param <V>   a type of values in a map.
-     * @return a {@link java.util.function.Function} .
+     * @param <K>   a type of keys in a map
+     * @param <V>   a type of values in a map
+     * @return a {@link java.util.function.Function}
      * @see java.util.Map#put(Object, Object)
      * @see java.util.function.Function
      * @since 0.1.0
@@ -107,9 +139,9 @@ public final class MapX {
      * all -&gt; map -&gt; map.putAll(all).</p>
      *
      * @param all a {@link java.util.Map} that is passed as the argument to {@link java.util.Map#putAll(Map)}
-     * @param <K> a type of keys in a map.
-     * @param <V> a type of values in a map.
-     * @return a {@link java.util.function.Consumer} .
+     * @param <K> a type of keys in a map
+     * @param <V> a type of values in a map
+     * @return a {@link java.util.function.Consumer}
      * @see java.util.Map#putAll(Map)
      * @see java.util.function.Consumer
      * @since 0.1.0
@@ -124,9 +156,9 @@ public final class MapX {
      *
      * @param key   a key of {@link K} that is passed as the first argument to {@link java.util.Map#put(Object, Object)}
      * @param value a value of {@link V} that is passed as the second argument to {@link java.util.Map#put(Object, Object)}
-     * @param <K>   a type of keys in a map.
-     * @param <V>   a type of values in a map.
-     * @return a {@link java.util.function.Consumer} .
+     * @param <K>   a type of keys in a map
+     * @param <V>   a type of values in a map
+     * @return a {@link java.util.function.Consumer}
      * @see java.util.Map#put(Object, Object)
      * @see java.util.function.Consumer
      * @since 0.1.0
@@ -140,9 +172,9 @@ public final class MapX {
      * key -&gt; map -&gt; map.remove(key).</p>
      *
      * @param key a key of {@link K} that is passed as the argument to {@link java.util.Map#remove(Object)}
-     * @param <K> a type of keys in a map.
-     * @param <V> a type of values in a map.
-     * @return a {@link java.util.function.Function} .
+     * @param <K> a type of keys in a map
+     * @param <V> a type of values in a map
+     * @return a {@link java.util.function.Function}
      * @see java.util.Map#remove(Object)
      * @see java.util.function.Function
      * @since 0.1.0
@@ -156,9 +188,9 @@ public final class MapX {
      * key -&gt; map -&gt; map.remove(key).</p>
      *
      * @param key a key of {@link K} that is passed as the argument to {@link java.util.Map#remove(Object)}
-     * @param <K> a type of keys in a map.
-     * @param <V> a type of values in a map.
-     * @return a {@link java.util.function.Consumer} .
+     * @param <K> a type of keys in a map
+     * @param <V> a type of values in a map
+     * @return a {@link java.util.function.Consumer}
      * @see java.util.Map#remove(Object)
      * @see java.util.function.Consumer
      * @since 0.1.0
@@ -172,14 +204,30 @@ public final class MapX {
      * other -&gt; map -&gt; map.equals(other).</p>
      *
      * @param other a {@link java.lang.Object} that is passed as the argument to {@link java.util.Map#equals(Object)}
-     * @param <K>   a type of keys in a map.
-     * @param <V>   a type of values in a map.
-     * @return a {@link java.util.function.Predicate} .
+     * @param <K>   a type of keys in a map
+     * @param <V>   a type of values in a map
+     * @return a {@link java.util.function.Predicate}
      * @see java.util.Map#equals(Object)
      * @see java.util.function.Predicate
      * @since 0.1.0
      */
     public static <K, V> Predicate<Map<? super K, ? super V>> equalsTo(Object other) {
         return map -> map.equals(other);
+    }
+
+    /**
+     * <p>Returns the carrying of {@link java.util.Map#equals(Object)} negative:
+     * other -&gt; map -&gt; !map.equals(other).</p>
+     *
+     * @param other a {@link java.lang.Object} that is passed as the argument to {@link java.util.Map#equals(Object)}
+     * @param <K>   a type of keys in a map
+     * @param <V>   a type of values in a map
+     * @return a {@link java.util.function.Predicate}
+     * @see java.util.Map#equals(Object)
+     * @see java.util.function.Predicate
+     * @since 0.1.0
+     */
+    public static <K, V> Predicate<Map<? super K, ? super V>> notEqualsTo(Object other) {
+        return map -> !map.equals(other);
     }
 }